+++ Sources/ZcashLightClientKit/Synchronizer/ClosureSDKSynchronizer.swift
@@ -194,6 +194,10 @@ extension ClosureSDKSynchronizer: ClosureSynchronizer {
         }
     }
 
+    public func refreshExchangeRateUSD() {
+        synchronizer.refreshExchangeRateUSD()
+    }
+
     /*
      It can be missleading that these two methods are returning Publisher even this protocol is closure based. Reason is that Synchronizer doesn't
      provide different implementations for these two methods. So Combine it is even here.